added outside click detection
