+++ app/frontend/src/components/Inputs/_Dropdown.scss
@@ -11,6 +11,7 @@
 }
 
 .dropdown-box {
+  background-color: $color-white;
   border: 1px solid $color-grey;
   border-radius: 8px;
   box-sizing: border-box;
